+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 Unreleased
 ==========
 
+- Modify the use of `prefix`. Now we use `s3://bucket/a/b/` instead of `-p a/b/`
+
 0.1.0 (2019-10-01)
 ==================